requirements.txt
annoy==1.17.1
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.3
Pillow==9.3.0
scipy==1.8.1
streamlit==1.11.1
tensorflow==2.11.0
tensorflow_hub==0.12.0
tqdm==4.64.0